Add JobDefinition class
