+++ python_modules/dagster/dagster/core/definitions/pipeline.py
@@ -551,68 +551,6 @@ class PipelineDefinition:
             "using an execution API function (e.g. `execute_pipeline`)."
         )
 
-    def execute_in_process(
-        self,
-        run_config: Optional[Dict[str, Any]] = None,
-        instance: Optional["DagsterInstance"] = None,
-        raise_on_error: bool = True,
-    ) -> "InProcessGraphResult":
-        """
-        (Experimental) Execute the "Job" (single mode pipeline) in-process, gathering results in-memory.
-
-        The executor_def on the Job will be ignored, and replaced with the in-process executor.
-        If using the default io_manager, it will switch from filesystem to in-memory.
-
-
-        Args:
-            run_config (Optional[Dict[str, Any]]:
-                The configuration for the run
-            instance (Optional[DagsterInstance]):
-                The instance to execute against, an ephemeral one will be used if none provided.
-
-        Returns:
-            InProcessGraphResult
-
-        """
-        from dagster.core.definitions.executor import execute_in_process_executor
-        from dagster.core.execution.execute_in_process import core_execute_in_process
-
-        run_config = check.opt_dict_param(run_config, "run_config")
-        check.invariant(
-            len(self._mode_definitions) == 1,
-            "execute_in_process only supported on job / single mode pipeline",
-        )
-
-        base_mode = self.get_mode_definition()
-        # create an ephemeral in process mode by replacing the executor_def and
-        # switching the default fs io_manager to in mem, if another was not set
-        in_proc_mode = ModeDefinition(
-            name="in_process",
-            executor_defs=[execute_in_process_executor],
-            resource_defs=swap_default_io_man(base_mode.resource_defs, self),
-            logger_defs=base_mode.loggers,
-            _config_mapping=base_mode.config_mapping,
-            _partitioned_config=base_mode.partitioned_config,
-        )
-
-        ephemeral_pipeline = PipelineDefinition(
-            name=self._name,
-            graph_def=self._graph_def,
-            mode_defs=[in_proc_mode],
-            hook_defs=self.hook_defs,
-            tags=self.tags,
-            version_strategy=self.version_strategy,
-        )
-
-        return core_execute_in_process(
-            node=self._graph_def,
-            ephemeral_pipeline=ephemeral_pipeline,
-            run_config=run_config,
-            instance=instance,
-            output_capturing_enabled=True,
-            raise_on_error=raise_on_error,
-        )
-
 
 class PipelineSubsetDefinition(PipelineDefinition):
     @property
@@ -660,6 +598,7 @@ def _get_pipeline_subset_def(
     Build a pipeline which is a subset of another pipeline.
     Only includes the solids which are in solids_to_execute.
     """
+    from .job import JobDefinition, JobSubsetDefinition
 
     check.inst_param(pipeline_def, "pipeline_def", PipelineDefinition)
     check.set_param(solids_to_execute, "solids_to_execute", of_type=str)
@@ -708,15 +647,27 @@ def _get_pipeline_subset_def(
                 )
 
     try:
-        sub_pipeline_def = PipelineSubsetDefinition(
-            name=pipeline_def.name,  # should we change the name for subsetted pipeline?
-            solid_defs=list({solid.definition for solid in solids}),
-            mode_defs=pipeline_def.mode_definitions,
-            dependencies=deps,
-            _parent_pipeline_def=pipeline_def,
-            tags=pipeline_def.tags,
-            hook_defs=pipeline_def.hook_defs,
-        )
+        if isinstance(pipeline_def, JobDefinition):
+            sub_pipeline_def: Union[
+                PipelineSubsetDefinition, JobSubsetDefinition
+            ] = JobSubsetDefinition(
+                name=pipeline_def.name,  # should we change the name for subsetted pipeline?
+                mode_def=pipeline_def.mode_definitions[0],
+                _parent_pipeline_def=pipeline_def,
+                tags=pipeline_def.tags,
+                hook_defs=pipeline_def.hook_defs,
+                graph_def=pipeline_def.graph,
+            )
+        else:
+            sub_pipeline_def = PipelineSubsetDefinition(
+                name=pipeline_def.name,  # should we change the name for subsetted pipeline?
+                solid_defs=list({solid.definition for solid in solids}),
+                mode_defs=pipeline_def.mode_definitions,
+                dependencies=deps,
+                _parent_pipeline_def=pipeline_def,
+                tags=pipeline_def.tags,
+                hook_defs=pipeline_def.hook_defs,
+            )
 
         return sub_pipeline_def
     except DagsterInvalidDefinitionError as exc:
@@ -1091,26 +1042,6 @@ def _create_run_config_schema(
         config_type_dict_by_key=config_type_dict_by_key,
         config_mapping=mode_definition.config_mapping,
     )
-
-
-def swap_default_io_man(resources: Dict[str, ResourceDefinition], job: PipelineDefinition):
-    """
-    Used to create the user facing experience of the default io_manager
-    switching to in-memory when using execute_in_process.
-    """
-    from dagster.core.storage.mem_io_manager import mem_io_manager
-    from .graph import default_job_io_manager
-
-    if (
-        # pylint: disable=comparison-with-callable
-        resources.get("io_manager") == default_job_io_manager
-        and job.version_strategy is None
-    ):
-        updated_resources = dict(resources)
-        updated_resources["io_manager"] = mem_io_manager
-        return updated_resources
-
-    return resources
 
 
 def _is_using_graph_job_op_apis(node_list: List[Node]):
